fix conflict and change directory
